Port non breaking changes from opin email setup
 - allow overwriting of render method
 - allow to configure a mixin for all emails (use in OPIN to change
   logo path and use SingleTemplateMixin for all emails)
